+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 2019-11-01 1.10.51
+# 2019-11-01 1.10.52
   - Socket Ping Interval increased to help app not disconnect as often
   - Don't start track if paused between tracks on wake
   - Log files: Name unnamed tracks sent to plotter (i.e.: DEAD_RECKON)
@@ -10,6 +10,7 @@
   - no save if network status has not changed (reduces extra log entries)
   - Software Update status
   - Download log file fix for files without dates
+  - Servo fault state uses fault_status instead of reason_unavailable, won't block UI
 
 # 2019-10-24 1.10.49
   - Download track images from Webcenter, create on table if unable